requirements.txt
Flask~=2.2.3
python-dotenv~=1.0.0
WTForms~=3.0.1
SQLAlchemy==2.0.6